This branch is an abandoned attempt at using DOM elements for icons
Pros:
- event listeners! Instead of this hacky bullshit: https://stackoverflow.com/questions/9395858/event-listener-on-a-css-pseudo-element-such-as-after-and-before

Cons:
- Every click on the page causes the inserted span(s) to disappear.
Possibly this would be solvable if I knew what event was triggering
that, and could put a "no, really, insert these elements" call in there.

I may have to come back and solve the Con given the Pro, but ... for now
I'm abandoning this in favor of the pseudoelement approach.